SVM Light Record Serialization
The SVMLightRecordReader was treating SVM Light feature
vectors as 0-based instead of 1-based. This was causing it
to read in incorrectly-sized vectors.

The unit test also added deserialized records to its
accumulator list multiple times and did not clear this list
between tests.

Also changed the variable name numRecordsAdded to the
more accurate numAttributesAdded.