Fix implementation of rootMeansSquaredError
In the previous implementation, because 1 and real.length are integers,
the result of ret in the first line will be that of an integer division,
which will be zero because real.length is always 1 or greater (if the
array is not empty - this also needs further fixing).

Anyway, the previous implementation essentially returns the root of the
sum of squared errors, not the root of the mean of the squared errors.

The fix divides the sum ret by real.length (hence computing the mean)
before the Math.sqrt().